requirements.txt
asgiref==3.5.2
boto3==1.26.39
botocore==1.29.39
Django==4.0.4
django-crispy-forms==1.14.0
django-storages==1.13.2
jmespath==1.0.1
Pillow==9.1.1
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.13
